requirements_copy.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.4.1
asyncpg==0.25.0
atlastk==0.13
click==8.0.3
fastapi==0.71.0
h11==0.12.0
idna==3.3
iso8601==0.1.16
pydantic==1.9.0
pypika-tortoise==0.1.3
pytz==2021.3
sniffio==1.2.0
starlette==0.17.1
tortoise-orm==0.18.1
typing_extensions==4.0.1
uuid==1.30
uvicorn==0.16.0
